% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{boxx_search}
\alias{boxx_search}
\alias{draw_search}
\title{Draw the current search() path in the terminal}
\usage{
boxx_search(labels = search(), width = NULL)

draw_search(labels = rev(search()), width = NULL)
}
\arguments{
\item{labels}{Character vector}

\item{width}{The common width of the boxes in number of characters.
If NULL, the width of the widest box will be used.}
}
\value{
Nothing.
}
\description{
Draw the current search() path in the terminal
}
\examples{
draw_search()

}
