% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_of.R
\name{environment_of}
\alias{environment_of}
\title{Get the environment of an object}
\usage{
environment_of(envir)
}
\arguments{
\item{envir}{An R object.}
}
\value{
A non-empty character string.
}
\description{
Get the environment of an object
}
\details{
Returns an \code{\link[base:environmentName]{environmentName(envir)}},
if \code{envir} is an \code{\link[base:environment]{environment}},
a function, or a formula.  In all other cases, NULL is returned.
}
\examples{
environment_of(globalenv())    ## R_GlobalEnv

environment_of(mean)           ## package:base

environment_of(abs)            ## package:base (primitive function)

environment_of(stats::median)  ## namespace:stats

environment_of(pi)             ## NULL

}
