% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_object.R
\name{find_object}
\alias{find_object}
\title{Find the environment where an object exists}
\usage{
find_object(name, mode = "any", from = parent.frame())
}
\arguments{
\item{name}{The name of the object to locate.}

\item{mode}{The [base::mode] of the object to locate.}

\item{from}{An [base::environment] to start search from.}
}
\value{
An [base::environment], or NULL.
}
\description{
Find the environment where an object exists
}
\details{
The object is looked for in environment `from`. If it is found there,
then `from` is returned.  If not found there, the parent environment
of `from` is searched, and so on, until the "empty" environment
([base::emptyenv()]) is reached. In such cases, no matching object
could be found and `NULL` is returned.
}
\examples{
find_object("pi")
find_object("pi", mode = "character")   ## non-existing
find_object("rnorm", mode = "function")

f <- local({
  a <- 42
  b <- 3.14
  function() a
})
env <- find_object("a", from = environment(f))
utils::ls.str(env)

}
