% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_object.R
\name{locate_object}
\alias{locate_object}
\title{Locate the original name and location of an object}
\usage{
locate_object(object, from = parent.frame())
}
\arguments{
\item{object}{The R object whose location should be identified.}

\item{from}{An \code{\link[base:environment]{environment}}, or an object
with an environment (e.g. a \code{\link[base:function]{function}} and a
\code{\link[base:tilde]{formula}}), to start search from.}
}
\value{
A named list with elements \code{name} and \code{envir}, where \code{name} is the
name of \code{object} as it is named in environment \code{envir}.
If the object could not be located when searching from environment \code{from},
then an error is produced.
}
\description{
Locate the original name and location of an object
}
\examples{
locate_object(pi)

a <- pi
locate_object(a)

a <- pi
locate_object(a, from = parent.env(globalenv()))

a <- local(pi)
locate_object(a)

}
