% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_fcn.R
\name{prune_fcn}
\alias{prune_fcn}
\title{Remove unused objects from a function's parent environments}
\usage{
prune_fcn(fcn, search = locate_object(fcn)$envir)
}
\arguments{
\item{fcn}{A \code{\link[base:function]{function}}.}

\item{search}{A \code{\link[base:environment]{environment}},
among the parents of \code{envir}, to be replaced.
It is possible to specify a list of alternative environments.}
}
\value{
A pruned version of \code{fcn}, with \code{prune_undo} attribute holding
an "undo" function. \emph{WARNING: Make sure to copy this attribute and then
remove it before exporting the function to an external process.}
}
\description{
Remove unused objects from a function's parent environments
}
