% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_function.R
\name{prune_function}
\alias{prune_function}
\title{Replace top parent environment of a function with a pruned environment}
\usage{
prune_function(object, populate = NULL, until = globalenv())
}
\arguments{
\item{object}{A \code{\link[base:function]{function}} or a
\code{\link[base:tilde]{formula}}) whose top parent environment
should be replaced with a pruned environment.}

\item{populate}{An optional, named list of objects to be assigned to
the top parent environment.}

\item{until}{An \code{\link[base:environment]{environment}} to consider
the last parent environment.  If `until` is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.}
}
\value{
A pruned version of `object`.
}
\description{
Replace top parent environment of a function with a pruned environment
}
\examples{
a <- 42
f <- local(function() a)

envs <- parent_envs(environment(f), until = environment())
names(envs)
f()

g <- prune_function(f, until = environment(), populate = list(a = 13))
envs <- parent_envs(environment(g), until = environment())
names(envs)
g()

## However, we also changed f() here!
f()

}
