% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_env.R
\name{top_env}
\alias{top_env}
\title{Find the top environments of a function or a formula}
\usage{
top_env(object, until = globalenv())
}
\arguments{
\item{object}{A \code{\link[base:function]{function}} or a
\code{\link[base:tilde]{formula}}) whose top parent environment
should be found.}

\item{until}{A \code{\link[base:environment]{environment}} to consider
the last parent environment.  If \code{until} is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.  It is also possible to specify a list of alternative
environments.}
}
\value{
An \code{\link[base:environment]{environment}}, which
is either the \code{until} environment or the empty environment.
}
\description{
The top environment is the top parent environment
}
\examples{
a <- 42
pi <- 3.14
f <- function() pi * a
env <- top_env(f)
print(env)
#stopifnot(identical(env, environment()))

f <- local({
  a <- 42
  local({
    pi <- 3.14
    function() pi * a
  })
})
env <- top_env(f)
print(env)
#stopifnot(identical(env, environment()))

make_fcn <- function() {
  a <- 42
  pi <- 3.14
  function() pi * a
}
f <- make_fcn()
env <- top_env(f)
print(env)
#stopifnot(identical(env, environment()))

}
\seealso{
\code{\link[=parent_env]{parent_env()}} and \code{\link[=parent_envs]{parent_envs()}}.
}
