% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_name.R
\name{environment_name}
\alias{environment_name}
\title{Get the name of an environment}
\usage{
environment_name(envir)
}
\arguments{
\item{envir}{An \code{\link[base:environment]{environment}}.}
}
\value{
A non-empty character string.
}
\description{
Get the name of an environment
}
\details{
This function returns what
\code{\link[base:environmentName]{environmentName(envir)}}, except when
it is empty, e.g. when \code{envir <- new.env()}.  In such cases, it returns
the hexadecimal string that is outputted by \code{print(envir)}.
}
\examples{
environment_name(emptyenv())
environment_name(globalenv())
environment_name(getNamespace("utils"))
environment_name(new.env())

}
