% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_of.R
\name{environment_of}
\alias{environment_of}
\title{Get the environment of an object}
\usage{
environment_of(envir)
}
\arguments{
\item{envir}{An R object.}
}
\value{
A non-empty character string.
}
\description{
Get the environment of an object
}
\details{
Returns an \code{\link[base:environment]{environment}}, if \code{envir} is
an environment, a function, or a formula.  In all other cases, NULL is
returned.
}
\examples{
## An environment
environment_of(globalenv())    ## <environment: R_GlobalEnv>

## A primitive function (always in the 'base' namespace)
environment_of(abs)            ## <environment: namespace:base>

## A package function
environment_of(mean)           ## <environment: namespace:base>

## Another package function
environment_of(stats::median)  ## <environment: namespace:stats>

## A formula
environment_of(y ~ x)          ## <environment: R_GlobalEnv>

## A non-function object
environment_of(pi)             ## NULL

}
