% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_object.R
\name{find_object}
\alias{find_object}
\title{Find the environment where an object exists}
\usage{
find_object(
  name = NULL,
  mode = "any",
  value = NULL,
  from = parent.frame(),
  until = emptyenv(),
  which = c("first", "last", "all")
)
}
\arguments{
\item{name}{The name of the object to locate.}

\item{mode}{The \code{\link[base:mode]{mode}} of the object to locate.}

\item{value}{The R object whose location should be identified.}

\item{from}{An \code{\link[base:environment]{environment}}, or an object
with an environment (e.g. a \code{\link[base:function]{function}} and a
\code{\link[base:tilde]{formula}}), to start search from.}

\item{until}{An \code{\link[base:environment]{environment}} to consider
the last parent environment.  If \code{until} is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.  It is also possible to specify a list of alternative
environments.}

\item{which}{If \code{"first"} or \code{"last"}, then the first or the last
occurance of \code{object} among the parent frames is identified and returned.
If \code{"all"}, then all occurances are returned.}
}
\value{
An \code{\link[base:environment]{environment}}, or NULL.
}
\description{
Find the environment where an object exists
}
\examples{
find_object(name = "pi")
find_object(name = "pi", mode = "character")   ## non-existing
find_object(name = "rnorm", mode = "function")

f <- local({
  a <- 42
  pi <- 3.14
  function() pi * a
})
env <- find_object(name = "a", from = f)
utils::ls.str(env)

f <- local({
  a <- 42
  local({
    pi <- 3.14
    function() pi * a
  })
})
env_a <- find_object(name = "a", from = f)
utils::ls.str(env_a)
env_pi <- find_object(name = "pi", from = f)
utils::ls.str(env_pi)
stopifnot(
  identical(environment(f), env_pi),
  identical(parent.env(env_pi), env_a)
)

find_object(value = rnorm)

my_sum <- rnorm
find_object(value = my_sum)

find_object(value = my_sum, from = parent_env())

my_sum <- local(sum)
find_object(value = my_sum)


my_find <- function(object, envir = parent.frame()) {
  find_object(value = object, from = envir)
}

my_fcn <- local({
  g <- sum  ## this is never found
  function(f = NULL) {
    if (is.null(f)) {
      f <- g  ## this is found first
      my_find(f)
    } else {
      my_find(f, envir = parent.frame()) # skip argument 'f'
    }
  }
})
my_fcn()
my_fcn(my_sum)

topenv()
}
