% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_object.R
\name{locate_object}
\alias{locate_object}
\title{Locate the original name and location of an object}
\usage{
locate_object(object, from = parent.frame())
}
\arguments{
\item{object}{The R object whose location should be identified.}

\item{from}{An \code{\link[base:environment]{environment}}, or an object
with an environment (e.g. a \code{\link[base:function]{function}} and a
\code{\link[base:tilde]{formula}}), to start search from.}
}
\value{
A named list with elements \code{name} and \code{envir}, where \code{name} is the
name of \code{object} as it is named in environment \code{envir}.
If the object could not be located when searching from environment
\code{from}, then NULL is returned.
}
\description{
Locate the original name and location of an object
}
\examples{
locate_object(rnorm)

my_sum <- rnorm
locate_object(my_sum)

locate_object(my_sum, from = parent.env(environment()))

my_sum <- local(sum)
locate_object(my_sum)


my_find <- function(object, envir = parent.frame()) {
  locate_object(object, from = envir)
}
my_fcn <- function() {
  my_sum <- sum
  my_find(my_sum)
}
my_fcn()

}
