% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_envs.R
\name{parent_envs}
\alias{parent_envs}
\title{Find all parent environment of an environment}
\usage{
parent_envs(envir)
}
\arguments{
\item{envir}{An [base::environment].}
}
\value{
A named list of environments, where the first element is `envir`
and the last is `emptyenv()`.
}
\description{
Find all parent environment of an environment
}
\examples{
parent_envs(emptyenv())
parent_envs(baseenv())
parent_envs(globalenv())
parent_envs(new.env(parent = baseenv()))

}
