% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_envs.R
\name{parent_envs}
\alias{parent_envs}
\title{Find all parent environments of an environment}
\usage{
parent_envs(envir, until = emptyenv(), extra = 0L)
}
\arguments{
\item{envir}{An \code{\link[base:environment]{environment}}
or an object with an environment, e.g. a function or a formula.}

\item{until}{A \code{\link[base:environment]{environment}} to consider
the last parent environment.  If \code{until} is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.  It is also possible to specify a list of alternative
environments.}

\item{extra}{Maximum number of additional parent environments to include
after a matching "until" one was identified.}
}
\value{
A named list of \code{\link[base:environment]{environment}}s, where
the first element is \code{envir} and the last is \code{until} or
\code{\link[base:emptyenv]{emptyenv()}}.
}
\description{
Find all parent environments of an environment
}
\examples{
parent_envs(emptyenv())
parent_envs(baseenv())
parent_envs(globalenv())
parent_envs(new.env(parent = baseenv()))

f <- local({
  a <- 42
  local({
    pi <- 3.14
    function() pi * a
  })
})
f_envs <- parent_envs(f, until = environment())
names(f_envs)

f_envs <- parent_envs(f, until = environment(), extra = 1L)
names(f_envs)

}
