% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_envs.R
\name{parent_envs}
\alias{parent_envs}
\title{Find all parent environments of an environment}
\usage{
parent_envs(envir, until = emptyenv())
}
\arguments{
\item{envir}{An \code{\link[base:environment]{environment}}.}

\item{until}{A \code{\link[base:environment]{environment}} to consider
the last parent environment.  If `until` is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.  It is also possible to specify a list of alternative
environments.}
}
\value{
A named list of \code{\link[base:environment]{environment}}s, where
the first element is `envir` and the last is `until` or
\code{\link[base:emptyenv]{emptyenv()}}.
}
\description{
Find all parent environments of an environment
}
\examples{
parent_envs(emptyenv())
parent_envs(baseenv())
parent_envs(globalenv())
parent_envs(new.env(parent = baseenv()))

f <- local({
  a <- 42
  local({
    pi <- 3.14
    function() pi * a
  })
})
parent_envs(environment(f), until = globalenv())

}
