% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_closure.R
\name{prune_function}
\alias{prune_function}
\title{Replace top parent environment of a function with a pruned environment}
\usage{
prune_function(object, until = globalenv())
}
\arguments{
\item{object}{A \code{\link[base:function]{function}} or a
\code{\link[base:tilde]{formula}}) whose top parent environment
should be replaced with a pruned environment.}

\item{until}{An \code{\link[base:environment]{environment}} to consider
the last parent environment.  If `until` is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.}
}
\value{
A pruned version of `object`.
}
\description{
Replace top parent environment of a function with a pruned environment
}
