% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_env.R
\name{replace_env}
\alias{replace_env}
\title{Replace one of the parent environments with another}
\usage{
replace_env(envir, search, replace, same_parent = TRUE)
}
\arguments{
\item{envir}{An \code{\link[base:environment]{environment}}.}

\item{search}{A \code{\link[base:environment]{environment}} to be
replaced environment with a pruned version.
It is also possible to specify a list of alternative environments.}

\item{replace}{A \code{\link[base:environment]{environment}}.}

\item{same_parent}{If TRUE, the parent environment of `replace` is
set to the parent environment of the replaced "search" environment.}
}
\value{
Invisibly, a named list parent environments until the one
that was replaced.
}
\description{
Replace one of the parent environments with another
}
\examples{
a <- 42
f <- local(function() a)

f_envs <- parent_envs(environment(f), until = environment(), extra = 1L)
names(f_envs)
y <- f()
y

new <- as.environment(list(a = 13, pi = 3.14))
old_parents <- replace_env(environment(f), search = environment(), replace = new)
names(old_parents)

f_envs <- parent_envs(environment(f), until = list(environment(), parent.env(environment())))
names(f_envs)

## Note that f() will now see a = 13 in the replaced environment
## rather than a = 42 in the calling environment
z <- f()
z

}
