% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_env.R
\name{top_env}
\alias{top_env}
\title{Find the top environments of a function}
\usage{
top_env(object, until = globalenv())
}
\arguments{
\item{object}{A \code{\link[base:function]{function}} or a
\code{\link[base:tilde]{formula}}) whose top parent environment
should be replaced with a pruned environment.}

\item{until}{A \code{\link[base:environment]{environment}} to consider
the last parent environment.  If `until` is not one of the parent
environments, then \code{\link[base:emptyenv]{emptyenv()}} is the
last one.  It is also possible to specify a list of alternative
environments.}
}
\value{
An \code{\link[base:environment]{environment}}.
}
\description{
Find the top environments of a function
}
\examples{
a <- 42
pi <- 3.14
f <- function() pi * a
env <- top_env(f)
print(env)
#stopifnot(identical(env, environment()))

f <- local({
  a <- 42
  local({
    pi <- 3.14
    function() pi * a
  })
})
env <- top_env(f)
print(env)
#stopifnot(identical(env, environment()))

make_fcn <- function() {
  a <- 42
  pi <- 3.14
  function() pi * a
}
f <- make_fcn()
env <- top_env(f)
print(env)
#stopifnot(identical(env, environment()))

}
